requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.85.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
black==26.3.1
blinker==1.9.0
boto3==1.42.70
botocore==1.42.70
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
coverage==7.13.5
cryptography==46.0.6
Deprecated==1.3.1
dictdiffer==0.9.0
distro==1.9.0
docstring_parser==0.17.0
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.8.0
groq==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.13.0
jmespath==1.1.0
librt==0.8.1
MarkupSafe==3.0.3
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
pluggy==1.6.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyGithub==2.5.0
Pygments==2.19.2
PyJWT==2.12.1
PyNaCl==1.6.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0
python-dotenv==1.2.2
pytokens==0.4.1
PyYAML==6.0.3
redis==7.4.0
regex==2026.2.28
requests==2.32.5
ruff==0.15.6
s3transfer==0.16.0
schedule==1.2.2
six==1.17.0
slack_sdk==3.41.0
sniffio==1.3.1
soupsieve==2.8.3
structlog==25.5.0
tavily==1.1.0
tavily-python==0.7.23
tenacity==9.1.4
tiktoken==0.12.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.7
wrapt==2.1.2
yarl==1.23.0
